+++ lib/timers.js
@@ -30,19 +30,20 @@ var TIMEOUT_MAX = 2147483647; // 2^31-1
 
 var debug = require('util').debuglog('timer');
 
-var asyncFlags = process._asyncFlags;
-var runAsyncQueue = process._runAsyncQueue;
-var loadAsyncQueue = process._loadAsyncQueue;
-var unloadAsyncQueue = process._unloadAsyncQueue;
+var tracing = require('tracing');
+var asyncFlags = tracing._asyncFlags;
+var runAsyncQueue = tracing._runAsyncQueue;
+var loadAsyncQueue = tracing._loadAsyncQueue;
+var unloadAsyncQueue = tracing._unloadAsyncQueue;
 
 // Same as in AsyncListener in env.h
 var kHasListener = 0;
 
 // Do a little housekeeping.
-delete process._asyncFlags;
-delete process._runAsyncQueue;
-delete process._loadAsyncQueue;
-delete process._unloadAsyncQueue;
+delete tracing._asyncFlags;
+delete tracing._runAsyncQueue;
+delete tracing._loadAsyncQueue;
+delete tracing._unloadAsyncQueue;
 
 
 // IDLE TIMEOUTS
