+++ test/simple/test-asynclistener-error-multiple-unhandled.js
@@ -21,6 +21,7 @@
 
 var common = require('../common');
 var assert = require('assert');
+var tracing = require('tracing');
 
 function onAsync0() {
   return 0;
@@ -45,8 +46,8 @@ var asyncNoHandleError1 = {
 };
 
 var listeners = [
-  process.addAsyncListener(asyncNoHandleError0),
-  process.addAsyncListener(asyncNoHandleError1)
+  tracing.addAsyncListener(asyncNoHandleError0),
+  tracing.addAsyncListener(asyncNoHandleError1)
 ];
 
 var uncaughtFired = false;
@@ -71,7 +72,7 @@ process.on('exit', function(code) {
 
   // Need to remove the async listeners or tests will always pass
   for (var i = 0; i < listeners.length; i++)
-    process.removeAsyncListener(listeners[i]);
+    tracing.removeAsyncListener(listeners[i]);
 
   assert.ok(uncaughtFired);
   console.log('ok');
