+++ test/simple/test-asynclistener-error.js
@@ -24,9 +24,10 @@ var assert = require('assert');
 var dns = require('dns');
 var fs = require('fs');
 var net = require('net');
+var tracing = require('tracing');
 
-var addListener = process.addAsyncListener;
-var removeListener = process.removeAsyncListener;
+var addListener = tracing.addAsyncListener;
+var removeListener = tracing.removeAsyncListener;
 var errorMsgs = [];
 var currentMsg = '';
 var caught = 0;
@@ -46,7 +47,7 @@ var callbacksObj = {
   }
 };
 
-var listener = process.createAsyncListener(callbacksObj);
+var listener = tracing.createAsyncListener(callbacksObj);
 
 process.on('exit', function(code) {
   removeListener(listener);
