+++ test/simple/test-asynclistener-remove-add-in-before.js
@@ -21,26 +21,27 @@
 
 var common = require('../common');
 var assert = require('assert');
+var tracing = require('tracing');
 var val;
 var callbacks = {
   create: function() {
     return 42;
   },
   before: function() {
-    process.removeAsyncListener(listener);
-    process.addAsyncListener(listener);
+    tracing.removeAsyncListener(listener);
+    tracing.addAsyncListener(listener);
   },
   after: function(context, storage) {
     val = storage;
   }
 };
 
-var listener = process.addAsyncListener(callbacks);
+var listener = tracing.addAsyncListener(callbacks);
 
 process.nextTick(function() {});
 
 process.on('exit', function(status) {
-  process.removeAsyncListener(listener);
+  tracing.removeAsyncListener(listener);
   assert.equal(status, 0);
   assert.equal(val, 42);
 });
