+++ test/simple/test-asynclistener-remove-in-before.js
@@ -21,22 +21,23 @@
 
 var common = require('../common');
 var assert = require('assert');
+var tracing = require('tracing');
 var done = false;
 var callbacks = {
   before: function() {
-    process.removeAsyncListener(listener);
+    tracing.removeAsyncListener(listener);
   },
   after: function() {
     done = true;
   }
 };
 
-var listener = process.addAsyncListener(callbacks);
+var listener = tracing.addAsyncListener(callbacks);
 
 process.nextTick(function() {});
 
 process.on('exit', function(status) {
-  process.removeAsyncListener(listener);
+  tracing.removeAsyncListener(listener);
   assert.equal(status, 0);
   assert.ok(done);
 });
