+++ test/simple/test-asynclistener-remove-inflight-error.js
@@ -21,6 +21,7 @@
 
 var common = require('../common');
 var assert = require('assert');
+var tracing = require('tracing');
 
 var set = 0;
 var asyncNoHandleError = {
@@ -29,13 +30,13 @@ var asyncNoHandleError = {
   }
 }
 
-var key = process.addAsyncListener(asyncNoHandleError);
+var key = tracing.addAsyncListener(asyncNoHandleError);
 
 process.nextTick(function() {
   throw 1;
 });
 
-process.removeAsyncListener(key);
+tracing.removeAsyncListener(key);
 
 var uncaughtFired = false;
 process.on('uncaughtException', function() {
