+++ test/simple/test-asynclistener.js
@@ -24,9 +24,10 @@ var assert = require('assert');
 var net = require('net');
 var fs = require('fs');
 var dgram = require('dgram');
+var tracing = require('tracing');
 
-var addListener = process.addAsyncListener;
-var removeListener = process.removeAsyncListener;
+var addListener = tracing.addAsyncListener;
+var removeListener = tracing.removeAsyncListener;
 var actualAsync = 0;
 var expectAsync = 0;
 
@@ -36,7 +37,7 @@ var callbacks = {
   }
 };
 
-var listener = process.createAsyncListener(callbacks);
+var listener = tracing.createAsyncListener(callbacks);
 
 process.on('exit', function() {
   process._rawDebug('expected', expectAsync);
